refactor: remove dead code detected by check_dead_code.py
Removed 17 unreachable symbols and their associated tests:

Source files:
- display.py: Deleted entirely (get_terminal_size, to_lines, to_blocks, format_diff)
- bytes.py: Encoding, MAX_DELETE_INTERVAL, RegionReplacementPatch, RegionReplacement, short_replacements, WHITESPACE, sort_whitespace
- clangdelta.py: clang_delta_works (moved to tests/helpers.py)
- genericlanguages.py: Substring class
- json.py: is_json function
- patching.py: ListPatches, Direction, Completed, LazyMutableRange, lazy_shuffle
- sequences.py: merged_intervals, with_deletions
- reducer.py: BasicReducer (moved to tests/helpers.py), RestartPass, SkipPass
- state.py: unused T TypeVar
- protocol.py: encode_bytes, decode_bytes
- worker.py: StdoutStream

Test utilities moved to tests/helpers.py:
- BasicReducer (used by reduce_with helper)
- clang_delta_works (used for pytest skipif markers)

Net reduction: 1137 lines removed.
