+++ tests/helpers.py
@@ -1,15 +1,84 @@
 import random
+import subprocess
 from collections.abc import Callable, Iterable
+from functools import lru_cache
 from typing import TypeVar
 
 import trio
+from attrs import define
 
-from shrinkray.passes.definitions import ReductionPass
+from shrinkray.passes.clangdelta import find_clang_delta
+from shrinkray.passes.definitions import ReductionPass, ReductionPump
 from shrinkray.passes.python import is_python
 from shrinkray.problem import BasicReductionProblem
-from shrinkray.reducer import BasicReducer, ShrinkRay
+from shrinkray.reducer import Reducer, ShrinkRay
 from shrinkray.state import sort_key_for_initial
 from shrinkray.work import WorkContext
+
+
+@define
+class BasicReducer[T](Reducer[T]):
+    """Simple reducer for testing - runs passes in sequence until no progress."""
+
+    reduction_passes: Iterable[ReductionPass[T]]
+    pumps: Iterable[ReductionPump[T]] = ()
+    _status: str = "Starting up"
+
+    def __attrs_post_init__(self) -> None:
+        self.reduction_passes = list(self.reduction_passes)
+
+    @property
+    def status(self) -> str:
+        return self._status
+
+    @status.setter
+    def status(self, value: str) -> None:
+        self._status = value
+
+    async def run_pass(self, rp: ReductionPass[T]) -> None:
+        await rp(self.target)
+
+    async def run(self) -> None:
+        await self.target.setup()
+
+        while True:
+            prev = self.target.current_test_case
+            for rp in self.reduction_passes:
+                self.status = f"Running reduction pass {rp.__name__}"
+                await self.run_pass(rp)
+            for pump in self.pumps:
+                self.status = f"Pumping with {pump.__name__}"
+                pumped = await pump(self.target)
+                if pumped != self.target.current_test_case:
+                    with self.backtrack(pumped):
+                        for rp in self.reduction_passes:
+                            self.status = f"Running reduction pass {rp.__name__} under pump {pump.__name__}"
+                            await self.run_pass(rp)
+            if prev == self.target.current_test_case:
+                return
+
+
+@lru_cache(maxsize=1)
+def clang_delta_works() -> bool:
+    """Check if clang_delta can actually execute.
+
+    This verifies not just that the binary exists, but that it can run.
+    On some systems (e.g., Ubuntu 24.04), creduce is installed but
+    clang_delta fails at runtime due to shared library issues.
+    """
+    clang_delta = find_clang_delta()
+    if not clang_delta:
+        return False
+    try:
+        # Run a simple test to verify clang_delta works
+        result = subprocess.run(
+            [clang_delta, "--help"],
+            capture_output=True,
+            timeout=5,
+        )
+        return result.returncode == 0
+    except (OSError, subprocess.TimeoutExpired):
+        return False
 
 
 T = TypeVar("T")
