+++ tests/test_byte_reduction_passes.py
@@ -1,12 +1,10 @@
 import ast
-from collections import Counter
 
 import pytest
-from hypothesis import assume, example, given
+from hypothesis import example, given
 from hypothesis import strategies as st
 
 from shrinkray.passes.bytes import (
-    WHITESPACE,
     ByteReplacement,
     debracket,
     find_ngram_endpoints,
@@ -14,10 +12,8 @@ from shrinkray.passes.bytes import (
     lower_bytes,
     lower_individual_bytes,
     short_deletions,
-    sort_whitespace,
 )
 from shrinkray.passes.patching import apply_patches
-from shrinkray.passes.python import is_python
 from shrinkray.problem import BasicReductionProblem, shortlex
 from shrinkray.work import WorkContext
 from tests.helpers import assert_reduces_to, direct_reductions, reduce_with
@@ -50,98 +46,6 @@ def test_short_deletions_can_delete_brackets() -> None:
 @given(st.binary())
 def test_ngram_endpoints(b):
     find_ngram_endpoints(b)
-
-
-def count_whitespace(b):
-    return len([c for c in b if c in WHITESPACE])
-
-
-def count_regions(b):
-    n = 0
-    is_whitespace_last = True
-    for c in b:
-        typ = c in WHITESPACE
-        if not typ and is_whitespace_last:
-            n += 1
-        is_whitespace_last = typ
-    return n
-
-
-@pytest.mark.skip
-@example(initial=b"\n0\r")
-@given(st.builds(bytes, st.lists(st.sampled_from(b"\t\n\r 0123."))))
-def test_sorting_whitespace(initial):
-    initial_count = count_whitespace(initial)
-
-    assume(initial_count > 0)
-
-    def is_interesting(tc):
-        assert count_whitespace(tc) == initial_count
-        assert Counter(tc) == Counter(initial)
-        return True
-
-    result = reduce_with([sort_whitespace], initial, is_interesting)
-
-    for i, c in enumerate(result):
-        assert (c in WHITESPACE) == (i < initial_count)
-
-
-@pytest.mark.skip
-@given(st.builds(bytes, st.lists(st.sampled_from(b"\t\n\r 0123."))))
-def test_sorting_whitespace_preserving_regions(initial):
-    initial_count = count_whitespace(initial)
-    initial_regions = count_regions(initial)
-
-    assume(initial_count > 0)
-    assume(initial_regions > 1)
-
-    def is_interesting(tc):
-        assert count_whitespace(tc) == initial_count
-        assert Counter(tc) == Counter(initial)
-        return count_regions(tc) == initial_regions
-
-    result = reduce_with([sort_whitespace], initial, is_interesting)
-
-    for run in runs_of_whitespace(result)[1:]:
-        assert len(run) <= 1
-
-
-@pytest.mark.skip
-@pytest.mark.parametrize(
-    "initial",
-    [
-        b"\t\t\t\t\nfrom\t\t\t\t\t\t\t\t\t\t.\timport\tA\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\nfrom\t\t\t.\t\t\t\t\t\t\t\t\t\timport\t\to\t\t\t\t\t\t\t\t\nfrom\t\t.\t\t\t\t\t\t\t\timport\t\tr\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t",
-        b"from\t\t.\t\t\t\t\t\t\t\t\timport\ta\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\nclass\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\ta\t\t\t\t\t\t\t\t\t\t\t(\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t)\t\t\t\t\t\t\t\t:\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t()\ndef\tr\t\t\t\t\t\t\t\t\t\t\t\t\t\t(\t\t\t\t\t\t\t\t\t\t\t\t\t\t)\t\t\t\t\t\t\t\t:\t\t\t\t\t\t\t\t\t\t\t\t\t\t...\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t",
-    ],
-)
-def test_sorting_whitespace_preserving_python(initial):
-    initial_count = count_whitespace(initial)
-
-    def is_interesting(tc):
-        assert count_whitespace(tc) == initial_count
-        assert Counter(tc) == Counter(initial)
-        return is_python(tc)
-
-    result = reduce_with([sort_whitespace], initial, is_interesting)
-
-    for run in runs_of_whitespace(result)[1:]:
-        assert len(run) <= 1
-
-
-def runs_of_whitespace(result):
-    whitespace_runs = []
-    i = 0
-    while i < len(result):
-        c = result[i]
-        if c not in WHITESPACE:
-            i += 1
-            continue
-        j = i + 1
-        while j < len(result) and result[j] in WHITESPACE:
-            j += 1
-        whitespace_runs.append(result[i:j])
-        i = j
-    return whitespace_runs
 
 
 def test_debracket():
